// src/components/dashboard/securitySettings/enum.ts
export enum SettingActionState {
	ACTIVE = "text-primary-400",
	INACTIVE = "text-error-main",
}

export enum Status {
	OFF = "OFF",
	TURN_ON = "TURN ON",
	VERIFY = "VERIFY",
	SETUP = "SETUP",
}

export enum IconTypes {
	TWO_FA = "TWO_FA",
	VERIFY_IDENTITY = "VERIFY_IDENTITY",
	ANTI_PHISHING = "ANTI_PHISHING",
	WITHDRAWAL_WHITELIST = "WITHDRAWAL_WHITELIST",
}